Actualizar onGroupParticipantsUpdate.js

+++ src/middlewares/onGroupParticipantsUpdate.js
@@ -1,47 +1,51 @@
 const { getProfileImageData } = require("../services/baileys");
 const fs = require("fs");
 const { onlyNumbers } = require("../utils");
-const { isActiveWelcomeGroup } = require("../utils/database");
+const {
+isActiveWelcomeGroup,
+getWelcomeMode
+} = require("../utils/database");
 const { warningLog } = require("../utils/logger");
 
-exports.onGroupParticipantsUpdate = async ({
-  groupParticipantsUpdate,
-  socket,
-}) => {
-  const remoteJid = groupParticipantsUpdate.id;
-  const userJid = groupParticipantsUpdate.participants[0];
+exports.onGroupParticipantsUpdate = async ({ groupParticipantsUpdate, socket }) => {
+const remoteJid = groupParticipantsUpdate.id;
+const userJid = groupParticipantsUpdate.participants[0];
 
-  if (!isActiveWelcomeGroup(remoteJid)) {
-    return;
+if (!isActiveWelcomeGroup(remoteJid)) {
+return;
+}
+
+const welcomeMode = getWelcomeMode(remoteJid);
+
+if (groupParticipantsUpdate.action === "add") {
+try {
+let buffer = null;
+let profileImage = null;
+
+  if (welcomeMode === "2") {
+    ({ buffer, profileImage } = await getProfileImageData(socket, userJid));
   }
 
-  if (groupParticipantsUpdate.action === "add") {
-    try {
-      const { buffer, profileImage } = await getProfileImageData(
-        socket,
-        userJid
-      );
+  const welcomeMessage = `¡𝗕𝗶𝗲𝗻𝘃𝗲𝗻𝗶𝗱@ 𝗮𝗹 𝗴𝗿𝘂𝗽𝗼! @${onlyNumbers(userJid)} 𝘗𝘳𝘦𝘴𝘦𝘯𝘵𝘢𝘯ᶜᵒⁿ 𝐟𝐨𝐭𝐨 y 𝐧𝐨𝐦𝐛𝐫𝐞 > Bot by Krampus OM Oᴘᴇʀᴀᴄɪᴏɴ Mᴀʀsʜᴀʟʟ ༴༎𝙾𝙼༎ > https://t.me/krampusiano`;
 
-      await socket.sendMessage(remoteJid, {
-        image: buffer,
-        caption: ` ¡𝗕𝗶𝗲𝗻𝘃𝗲𝗻𝗶𝗱@ 𝗮𝗹 𝗴𝗿𝘂𝗽𝗼!
-@${onlyNumbers(userJid)}
-𝘗𝘳𝘦𝘴𝘦𝘯𝘵𝘢𝘵𝘦 ᶜᵒⁿ 𝐟𝐨𝐭𝐨 y 𝐧𝐨𝐦𝐛𝐫𝐞 
-
-
-> Bot by Krampus OM
-Oᴘᴇʀᴀᴄɪᴏɴ Mᴀʀsʜᴀʟʟ ༴༎𝙾𝙼༎
-> https://t.me/krampusiano`,
-        mentions: [userJid],
-      });
-
-      if (!profileImage.includes("default-user")) {
-        fs.unlinkSync(profileImage);
-      }
-    } catch (error) {
-      warningLog(
-        "👻 𝙺𝚛𝚊𝚖𝚙𝚞𝚜.𝚋𝚘𝚝 👻 No se pudo enviar el mensaje de Bienvenida"
-      );
+  if (welcomeMode === "2") {
+    await socket.sendMessage(remoteJid, {
+      image: buffer,
+      caption: welcomeMessage,
+      mentions: [userJid],
+    });
+    if (!profileImage.includes("default-user")) {
+      fs.unlinkSync(profileImage);
     }
+  } else {
+    await socket.sendMessage(remoteJid, {
+      text: welcomeMessage,
+      mentions: [userJid],
+    });
   }
+} catch (error) {
+  warningLog("👻 𝙺𝚛𝚊𝚖𝚙𝚞𝚜.𝚋𝚘𝚝 👻 No se pudo enviar el mensaje de Bienvenida");
+}
+}
 };
+
